pyproject.toml
[tool.poetry]
name = "picobot"
version = "0.1.0"
description = ""
authors = ["Arthur Mesquita Pickcius <arthur.pickcius@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.2"
python-telegram-bot = "^13.11"
dataset = "^1.1"
pillow = "^9.0"
python-slugify = "^4.0"
emoji = "^0.6.0"
typed-ast = "^1.5.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"
summon-python = {extras = ["core"], version = "^0.1.3"}
pre-commit = "^2.19.0"
flake8-bugbear = "^22.4.25"
mypy = "^0.961"
types-emoji = "^1.7.1"
types-Pillow = "^9.0.19"
types-python-slugify = "^5.0.4"
pylint = "^2.14.2"

[tool.black]
skip_string_normalization = true
line_length = 100

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

[tool.summon.plugins.python]
test-modules = ["tests"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
